fix: salesForecastPercentage Double로 변경

+++ growthmate-core/src/main/java/com/growup/growthmate/company/domain/CompanyComparison.java
@@ -24,5 +24,5 @@ public class CompanyComparison {
     private Long salesForecast;
 
     @Column(nullable = false)
-    private Integer salesForecastPercentage;
+    private Double salesForecastPercentage;
 }
