+++ growthmate-core/src/main/resources/db/migration/V7__company-comparision.sql
@@ -3,7 +3,7 @@ CREATE TABLE company_comparison
     company_comparision_id    BIGINT AUTO_INCREMENT NOT NULL,
     company_id                BIGINT                NOT NULL,
     sales_forecast            BIGINT                NOT NULL,
-    sales_forecast_percentage INT                NOT NULL,
+    sales_forecast_percentage DOUBLE                NOT NULL,
     CONSTRAINT pk_companycomparison PRIMARY KEY (company_comparision_id)
 );
